requirements.txt
# Python 3.10+

# 🔥 PyTorch
torch>=2.10.0
torchvision>=0.20.0
torchaudio>=2.10.0

# 📦 Core libraries
numpy>=1.24,<2.0
opencv-python>=4.8
matplotlib>=3.7
tqdm>=4.65

# 📊 Detection / evaluation
pycocotools>=2.0.6

# ⚙️ Config management
hydra-core>=1.3
omegaconf>=2.3
PyYAML>=6.0

# 🤖 YOLO (newer needed for CUDA 13 compatibility)
ultralytics>=8.4.0

# Sliced inference (EXP-003 SAHI + Ultralytics)
sahi>=0.11.14

# Optional: EXP-A004 ANTS `roi_mode: dbscan` in configs/expA004_ants_v1.yaml
# scikit-learn>=1.3.0
